backend/requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
python-multipart
sqlalchemy
passlib[bcrypt]
pyjwt
tensorflow
